Ivan's Restructure

+++ QuestUp/Assets/[My Assets]/Scriptable Objects/[Scripts]/LevelSO.cs
@@ -5,6 +5,6 @@ namespace QuestUp
     [CreateAssetMenu(fileName = "Level", menuName = "ScriptableObjects/LevelSO")]
     public class LevelSO : ScriptableObject
     {
-        [field: SerializeField] public float LevelDuration { get; private set; }
+        [field: SerializeField] public int LevelDuration { get; private set; }
     }
 }
